sdawe

+++ pages/bossPage.php
@@ -85,7 +85,7 @@ echo "Welcome user " . $userID;
             "SELECT order_id , orderDate  , item , roomNumber, problem , userID , dataStatusId ,money_order FROM workorder WHERE  dataStatusId = 1 AND money_order >=0 AND orderStatus IS NULL "
         );
         if (mysqli_num_rows($result) > 0) {
-            ?>
+        ?>
             <table border="1">
                 <tr>
                     <td>Захиалгын код</td>
@@ -102,7 +102,7 @@ echo "Welcome user " . $userID;
                 <?php
                 $i = 0;
                 while ($row = mysqli_fetch_array($result)) {
-                    ?>
+                ?>
                     <tr>
                         <td>
                             <?php echo $row["order_id"]; ?>
@@ -127,38 +127,35 @@ echo "Welcome user " . $userID;
                         </td>
                         <td>
                             <?php if ($row['dataStatusId'] == 1) {
-                                ?>
+                            ?>
                                 <p class="status">Баталсан </p>
-                                <?php
+                            <?php
                             } else {
-                                ?>
+                            ?>
                                 <p class="status2">Батлаагүй</p>
-                                <?php
+                            <?php
 
-                            }
-                            ;
+                            };
                             ?>
                         </td>
                         <td>
                             <button>
-                                <a onclick="return confirm ('Та төсөв батлахдаа итгэлтэй байна уу  ? ')"
-                                    href="../service/uproveBudget.php?id=<?php echo $row["order_id"]; ?>">Батлах</a>
+                                <a onclick="return confirm ('Та төсөв батлахдаа итгэлтэй байна уу  ? ')" href="../service/uproveBudget.php?id=<?php echo $row["order_id"]; ?>">Батлах</a>
                             </button>
                         </td>
                         <td>
                             <button>
-                                <a onclick="return confirm ('Та захиалгыг устгахдаа итгэлтэй байна уу  ? ')"
-                                    href="userOrderDelete.php?id=<?php echo $row["order_id"]; ?>">Устгах</a> </button>
+                                <a onclick="return confirm ('Та захиалгыг устгахдаа итгэлтэй байна уу  ? ')" href="userOrderDelete.php?id=<?php echo $row["order_id"]; ?>">Устгах</a> </button>
                         </td>
                     </tr>
 
-                    <?php
+                <?php
                     $i++;
                 }
                 ?>
             </table>
 
-            <?php
+        <?php
         } else {
             echo "Өгөгдөл байхгүй";
         }
@@ -173,15 +170,16 @@ echo "Welcome user " . $userID;
         <?php
         $result = mysqli_query(
             $conn,
-            "SELECT order_id , orderDate  , item , problem , userID , dataStatusId ,money_order FROM workorder WHERE  dataStatusId = 1 AND money_order >=0 AND orderStatus = 1  "
+            "SELECT * FROM workorder WHERE  dataStatusId = 1 AND money_order >=0 AND orderStatus = 1  "
         );
         if (mysqli_num_rows($result) > 0) {
-            ?>
+        ?>
             <table border="1">
                 <tr>
                     <td>Захиалгын код</td>
                     <td>Ажилтны код</td>
                     <td>Захиалга өгсөн өдөр</td>
+                    <td>Тоот</td>
                     <td>Эд хөрөнгийн нэр </td>
                     <td>Асуудал </td>
                     <td>Үнийн санал </td>
@@ -192,7 +190,7 @@ echo "Welcome user " . $userID;
                 <?php
                 $i = 0;
                 while ($row = mysqli_fetch_array($result)) {
-                    ?>
+                ?>
                     <tr>
                         <td>
                             <?php echo $row["order_id"]; ?>
@@ -202,6 +200,9 @@ echo "Welcome user " . $userID;
                         </td>
                         <td>
                             <?php echo $row["orderDate"]; ?>
+                        </td>
+                        <td>
+                            <?php echo $row["roomNumber"]; ?>
                         </td>
                         <td>
                             <?php echo $row["item"]; ?>
@@ -214,29 +215,28 @@ echo "Welcome user " . $userID;
                         </td>
                         <td>
                             <?php if ($row['dataStatusId'] == 1) {
-                                ?>
+                            ?>
                                 <p class="status">Баталсан </p>
-                                <?php
+                            <?php
                             } else {
-                                ?>
+                            ?>
                                 <p class="status2">Батлаагүй</p>
-                                <?php
+                            <?php
 
-                            }
-                            ;
+                            };
                             ?>
                         </td>
-                        
-                      
+
+
                     </tr>
 
-                    <?php
+                <?php
                     $i++;
                 }
                 ?>
             </table>
 
-            <?php
+        <?php
         } else {
             echo "Өгөгдөл байхгүй";
         }
